refactor(commerce_pos_gratuity): move the add gratuity button to payment page
This was much trickier than having it be on the sale form, because the payment form doesn't have as many thigns built in for handling the form and order rebuild. Took me a while to track down how that works. Also, the buttons are not added in the same way, so I had to refactor that a little bit.

https://www.drupal.org/node/2857482